backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
supabase==1.2.0
